back/requirements.txt
flask
flask-cors
pytest
PyPDF2
pandas
python-dotenv